requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.9.3
astor==0.8.1
astunparse==1.6.3
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.1.1
cloudpickle==2.1.0
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
dm-tree==0.1.7
fonttools==4.36.0
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==2.10.0
hydra-core==1.3.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.12.0
install==1.3.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.3
mdurl==0.1.2
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
# Editable install with no version control (pinnstf2==0.1.1)
-e f:\postdoc\research\18-yaw_misalignment\dtu_les_pinn\pinns-tf2-main
plotting==0.0.7
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDOE==0.3.8
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.2.1
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.7.1
rootutils==1.0.7
rsa==4.9
scipy==1.4.1
seaborn==0.11.2
six==1.16.0
tensorboard==2.1.1
tensorboard-plugin-wit==1.8.1
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
tensorflow-probability==0.9.0
termcolor==1.1.0
tqdm==4.66.4
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
xarray==0.20.2
zipp==3.8.1
